+++ app.js
@@ -90,6 +90,7 @@ app.use(
         "'self'",
         'https://api.mapbox.com',
         'https://events.mapbox.com',
+        'ws://127.0.0.1:63815/',
       ],
     },
   }),
