+++ public/js/index.js
@@ -1,0 +1,23 @@
+/* eslint-disable */
+import '@babel/polyfill';
+import { displayMap } from './mapbox';
+import { login } from './login';
+
+// DOM ELEMENTS
+const mapContainer = document.getElementById('map');
+const loginForm = document.querySelector('.form');
+
+// DELEGATION
+if (mapContainer) {
+  const locations = JSON.parse(mapContainer.dataset.locations);
+  displayMap(locations);
+}
+
+if (loginForm) {
+  loginForm.addEventListener('submit', (e) => {
+    e.preventDefault();
+    const email = document.getElementById('email').value;
+    const password = document.getElementById('password').value;
+    login(email, password);
+  });
+}
