+++ public/js/login.js
@@ -1,7 +1,8 @@
 /*eslint-disable */
-// import axios from 'axios';
+import axios from 'axios';
+import { showAlert } from './alert';
 
-const login = async (email, password) => {
+export const login = async (email, password) => {
   try {
     const res = await axios({
       method: 'POST',
@@ -11,20 +12,15 @@ const login = async (email, password) => {
         password,
       },
     });
+
     if (res.data.status === 'success') {
-      alert('Login Successful');
+      showAlert('sucess', 'Login Successful');
+
       window.setTimeout(() => {
         location.assign('/');
       }, 1500);
     }
   } catch (err) {
-    alert(err.response.data.message, 'Login Failed');
+    showAlert('error', err.response.data.message);
   }
 };
-
-document.querySelector('.form').addEventListener('submit', (e) => {
-  e.preventDefault();
-  const email = document.getElementById('email').value;
-  const password = document.getElementById('password').value;
-  login(email, password);
-});
