+++ views/base.pug
@@ -21,6 +21,7 @@ html
     include _footer
 
     // SCRIPTS
-    script(src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.18.0/axios.min.js")
-    script(src='/js/mapbox.js')
-    script(src='/js/login.js')
+    //- script(src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.18.0/axios.min.js")
+    //- script(src='/js/mapbox.js')
+    //- script(src='/js/login.js')
+    script(src="/js/bundle.js")
